HHH-11764 - Fix PGGeometryTypeDescriptor
