+++ hibernate-spatial/src/test/resources/hibernate.properties
@@ -28,9 +28,9 @@ hibernate.connection.password @jdbc.pass@
 
 ## Configs for spatial databases (used during testing on local dev environment).
 #
-#hibernate.dialect org.hibernate.spatial.dialect.postgis.PostgisDialect
+#hibernate.dialect org.hibernate.spatial.dialect.postgis.PostgisPG95Dialect
 #hibernate.connection.driver_class org.postgresql.Driver
-#hibernate.connection.url jdbc:postgresql://localhost:5432:hibbrtru
+#hibernate.connection.url jdbc:postgresql://localhost:5432/hibbrtru
 #hibernate.connection.username hibbrtru
 #hibernate.connection.password hibbrtru
 
@@ -84,7 +84,7 @@ hibernate.connection.password @jdbc.pass@
 ##
 ## MySQL 5 InnoDDB dialect
 ##
-#hibernate.dialect org.hibernate.spatial.dialect.mysql.MySQLSpatial5InnoDBDialect
+#hibernate.dialect org.hibernate.spatial.dialect.mysql.MySQL56SpatialDialect
 #hibernate.connection.driver_class com.mysql.jdbc.Driver
 #hibernate.connection.url jdbc:mysql://localhost/testhbs
 #hibernate.connection.username hibernate
